% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data_functions.R
\name{generate_test_data}
\alias{generate_test_data}
\title{Generate multi-dimensional test data for spline smoothing.}
\usage{
generate_test_data(n, P, split = 0.8)
}
\arguments{
\item{n}{Numer of samples}

\item{P}{Spatial dimension}

\item{split}{A value between \code{0} and \code{1} for the train / test split.}
}
\value{
A list of the covarite matrices for the train and test data \code{X_train} and \code{X_test} and of the variable of interest \code{y_train} and \code{y_test}.
}
\description{
Generate a \code{P}-dimensional test data set based on a sigmoid function.
}
\examples{
generate_test_data(100, 2)

}
