% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgss_main.R
\name{predict_smooth}
\alias{predict_smooth}
\title{Predictions from model}
\usage{
predict_smooth(model_smooth, X)
}
\arguments{
\item{model_smooth}{A spline model resulting from \code{CG_smooth}, \code{PCG_smooth}, or \code{MGCG_smooth}.}

\item{X}{Matrix containing the new observations.}
}
\value{
Vector of length \code{nrow(X)} of predictions.
}
\description{
Makes predictions of new observations from a fitted spline model.
}
